pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.27.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.5.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
